Rename device selector
