+++ ui/src/components/DeviceSelectorHostname.vue
